+++ src/index.js
@@ -69,17 +69,15 @@ function attachAfterEmitHook(compiler, callback) {
   }
 }
 
-/** Convenience function to add a webpack compilation error. */
-function addCompilationError(compilation, message) {
-  compilation.errors.push(`Sentry CLI Plugin: ${message}`);
-}
-
 class SentryCliPlugin {
   constructor(options = {}) {
-    this.debug = options.debug || false;
+    const defaults = {
+      debug: false,
+      finalize: true,
+      rewrite: true,
+    };
 
-    // By default we want that rewrite is true
-    this.options = Object.assign({ rewrite: true }, options);
+    this.options = Object.assign({}, defaults, options);
 
     if (options.include) this.options.include = toArray(options.include);
     if (options.ignore) this.options.ignore = toArray(options.ignore);
@@ -114,6 +112,7 @@ class SentryCliPlugin {
     }
   }
 
+  /** Returns whether this plugin should emit any data to stdout. */
   isSilent() {
     return this.options.silent === true;
   }
@@ -151,6 +150,10 @@ class SentryCliPlugin {
             this.outputDebug('Finalizing release:\n', release);
             return Promise.resolve(release);
           },
+          setCommits: (release, config) => {
+            this.outputDebug('Calling set-commits with:\n', config);
+            return Promise.resolve(release, config);
+          },
         },
       };
     }
@@ -161,14 +164,17 @@ class SentryCliPlugin {
   /**
    * Returns a Promise that will solve to the configured release.
    *
-   * If no release is specified, it uses Sentry CLI to propose a version. The
-   * release string is always trimmed.
+   * If no release is specified, it uses Sentry CLI to propose a version.
+   * The release string is always trimmed.
+   * Returns undefined if proposeVersion failed.
    */
   getReleasePromise() {
     return (this.options.release
       ? Promise.resolve(this.options.release)
       : this.cli.releases.proposeVersion()
-    ).then(version => `${version}`.trim());
+    )
+      .then(version => `${version}`.trim())
+      .catch(() => undefined);
   }
 
   /** Checks if the given named entry point should be handled. */
@@ -300,26 +306,54 @@ class SentryCliPlugin {
   finalizeRelease(compilation) {
     const {
       include,
-      errorHandler = (err, invokeErr) => {
+      errorHandler = (_, invokeErr) => {
         invokeErr();
       },
     } = this.options;
-
-    if (!include) {
-      addCompilationError(compilation, '`include` option is required');
-      return Promise.resolve();
-    }
 
     let release;
     return this.release
       .then(proposedVersion => {
         release = proposedVersion;
+
+        if (!include) {
+          throw new Error(`\`include\` option is required`);
+        }
+
+        if (!release) {
+          throw new Error(
+            `Unabled to determine version. Make sure to include \`release\` option or use the environment that supports auto-detection https://docs.sentry.io/cli/releases/#creating-releases`
+          );
+        }
+
         return this.cli.releases.new(release);
       })
       .then(() => this.cli.releases.uploadSourceMaps(release, this.options))
-      .then(() => this.cli.releases.finalize(release))
+      .then(() => {
+        const { commit, previousCommit, repo, auto } =
+          this.options.setCommits || this.options;
+
+        if (auto || (repo && commit)) {
+          this.cli.releases.setCommits(release, {
+            commit,
+            previousCommit,
+            repo,
+            auto,
+          });
+        }
+      })
+      .then(() => {
+        if (this.options.finalize) {
+          return this.cli.releases.finalize(release);
+        }
+        return undefined;
+      })
       .catch(err =>
-        errorHandler(err, () => addCompilationError(compilation, err.message))
+        errorHandler(
+          err,
+          () => compilation.errors.push(`Sentry CLI Plugin: ${err.message}`),
+          compilation
+        )
       );
   }
 
@@ -328,7 +362,7 @@ class SentryCliPlugin {
     const compilerOptions = compiler.options || {};
     ensure(compilerOptions, 'module', Object);
 
-    if (this.debug) {
+    if (this.options.debug) {
       this.injectReleaseWithDebug(compilerOptions);
     } else {
       this.injectRelease(compilerOptions);