+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## v1.9.3
+
+- chore: Bump sentry-cli to `1.49.0`
+- fix: Dont fail compilation if there is no release available (#155)
+- fix: Update auto/repo logic for `setCommit` option (#156)
+
+## v1.9.2
+
+- chore: Resolve Snyk as dependency issues (#152)
+
+## v1.9.1
+
+- ref: Allow for nested setCommits (#142)
+- fix: Fixed TS definitions export error (#145)
+
+## v1.9.0
+
+- feat: Add `setCommits` options (#139)
+- chore: Add `TypeScript` definition file (#137)
+- meta: Bump sentry-cli to `1.48.0`
+
 ## v1.8.1
 
 - meta: Bump sentry-cli to `1.47.1`